requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
Flask==3.0.3
h3==4.1.2
interchange==2021.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
jplephem==2.22
MarkupSafe==3.0.2
monotonic==1.6
neo4j==5.25.0
numpy==2.1.3
packaging==24.1
pansi==2020.7.3
py2neo==2021.2.4
pycparser==2.22
Pygments==2.18.0
pyswisseph==2.10.3.2
pytz==2024.2
PyYAML==6.0.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sgp4==2.23
six==1.16.0
skyfield==1.49
timezonefinder==6.5.7
urllib3==2.2.3
Werkzeug==3.0.4
python-dotenv==1.0.1
